prepare for "rewrite"
