+++ core/src/main/kotlin/net/bladehunt/window/core/primitive/AbstractRow.kt
@@ -21,16 +21,18 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package net.bladehunt.window.core.layout
+package net.bladehunt.window.core.primitive
 
 import kotlin.math.max
 import kotlin.math.roundToInt
 import net.bladehunt.window.core.Context
 import net.bladehunt.window.core.layer.OffsetLimitedLayer
+import net.bladehunt.window.core.layout.Window
 import net.bladehunt.window.core.util.IntPair
 import net.bladehunt.window.core.util.Size
+import net.bladehunt.window.core.widget.Layout
 
-abstract class AbstractRow<T>(size: Size = Size(0, 0)) : LayoutWidget<T>(size) {
+abstract class AbstractRow<T>(size: Size = Size(0, 0)) : Layout<T>(size) {
     override fun calculateSize(node: Window.Node<T>, context: Context): Size {
         val (sizeX, sizeY, flexX, flexY) = size
