--- minestom/widgets/src/main/kotlin/net/bladehunt/minestom/widgets/dsl/AbstractWidgetDsl.kt
@@ -1,45 +1,0 @@
-/*
- * Copyright 2024 BladehuntMC
- * Copyright 2024 oglassdev
- *
- * Permission is hereby granted, free of charge, to any person obtaining a
- * copy of this software and associated documentation files (the “Software”),
- * to deal in the Software without restriction, including without limitation the
- * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
- * sell copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
- * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
- * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
- * OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package net.bladehunt.minestom.widgets.dsl
-
-import net.bladehunt.minestom.widgets.core.*
-import net.bladehunt.window.core.Modifier
-import net.bladehunt.window.core.WindowDsl
-import net.bladehunt.window.core.util.Size
-
-@WindowDsl
-inline fun column(size: Size = Size(), modifier: Modifier<Column>): Column =
-    Column(size).apply(modifier)
-
-@WindowDsl
-inline fun row(size: Size = Size(), modifier: Modifier<Row>): Row = Row(size).apply(modifier)
-
-@WindowDsl
-inline fun auto(size: Size = Size(), modifier: Modifier<Auto>): Auto = Auto(size).apply(modifier)
-
-@WindowDsl
-inline fun container(size: Size = Size(), modifier: Modifier<Container>): Container =
-    Container(size).apply(modifier)
-
-@WindowDsl inline fun button(modifier: Modifier<Button>): Button = Button().apply(modifier)
